+++ apps/docs/.env.example
@@ -22,7 +22,8 @@ AUI_ANONYMOUS_SESSION_SECRET=
 # Daily request ceilings applied in addition to the short-window IP rate limit.
 # AUI_IP_REQUESTS_PER_DAY=200
 # AUI_IDENTITY_REQUESTS_PER_DAY=50
-# AUI_GLOBAL_REQUESTS_PER_DAY=5000
+# Required outside development. Set from measured peak request volume and spend.
+AUI_GLOBAL_REQUESTS_PER_DAY=
 # AUI_ANONYMOUS_SESSIONS_PER_IP_PER_DAY=100
 # AUI_DOWNLOAD_REQUESTS_PER_DAY=20
 # Optional only behind a trusted reverse proxy that overwrites this header.
