+++ apps/docs/app/sign-up/[[...sign-up]]/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { SignUp } from "@clerk/nextjs";
 import { AI_BUILDER_RETURN_PATH } from "@/lib/ai-builder-routes";
