--- apps/docs/components/anonymous-session-bootstrap.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import { useEffect } from "react";
-import { ensureAnonymousSession } from "@/lib/anonymous-session-client";
-
-export function AnonymousSessionBootstrap() {
-  useEffect(() => {
-    void ensureAnonymousSession().catch(() => {});
-  }, []);
-
-  return null;
-}
