+++ apps/docs/lib/rate-limit.ts
@@ -9,6 +9,15 @@ export function positiveSafeInteger(
   return Number.isSafeInteger(parsed) && parsed > 0 ? parsed : fallback;
 }
 
+function getGlobalDailyRequestLimit(): number {
+  const limit = positiveSafeInteger(process.env.AUI_GLOBAL_REQUESTS_PER_DAY, 0);
+  if (limit > 0) return limit;
+  if (process.env.NODE_ENV === "test") return 5_000;
+  throw new Error(
+    "AUI_GLOBAL_REQUESTS_PER_DAY must be set to a positive safe integer.",
+  );
+}
+
 const getRatelimits = async () => {
   if (isDev) return null;
   const { Redis } = await import("@upstash/redis");
@@ -39,10 +48,7 @@ const getRatelimits = async () => {
     inferenceGlobal: new Ratelimit({
       redis,
       prefix: "aui:inference:global",
-      limiter: Ratelimit.fixedWindow(
-        positiveSafeInteger(process.env.AUI_GLOBAL_REQUESTS_PER_DAY, 5_000),
-        "1d",
-      ),
+      limiter: Ratelimit.fixedWindow(getGlobalDailyRequestLimit(), "1d"),
     }),
     sessionIssuanceBurst: new Ratelimit({
       redis,
@@ -77,6 +83,30 @@ const getRatelimits = async () => {
 };
 
 const ratelimitsPromise = getRatelimits();
+type Ratelimits = NonNullable<Awaited<typeof ratelimitsPromise>>;
+
+async function runRateLimitCheck(
+  req: Request,
+  check: (ratelimits: Ratelimits) => Promise<Response | null>,
+): Promise<Response | null> {
+  try {
+    const ratelimits = await ratelimitsPromise;
+    if (!ratelimits) return null;
+    return await check(ratelimits);
+  } catch (error) {
+    console.error(
+      JSON.stringify({
+        level: "error",
+        message: "rate_limit_unavailable",
+        requestId: req.headers.get("x-vercel-id"),
+        error: error instanceof Error ? error.message : String(error),
+      }),
+    );
+    return new Response("Rate limiting temporarily unavailable", {
+      status: 503,
+    });
+  }
+}
 
 function firstIp(value: string | null): string | null {
   return value?.split(",")[0]?.trim() || null;
@@ -102,64 +132,71 @@ export async function checkRateLimit(
   req: Request,
   identity?: string,
 ): Promise<Response | null> {
-  const ratelimits = await ratelimitsPromise;
-  if (!ratelimits) return null;
+  return runRateLimitCheck(req, async (ratelimits) => {
+    const ip = getClientIp(req);
+    const burstResult = await ratelimits.inferenceIpBurst.limit(ip);
+    if (!burstResult.success) {
+      return new Response("Rate limit exceeded", { status: 429 });
+    }
 
-  const ip = getClientIp(req);
-  const burstResult = await ratelimits.inferenceIpBurst.limit(ip);
-  if (!burstResult.success) {
-    return new Response("Rate limit exceeded", { status: 429 });
-  }
+    const dailyIpResult = await ratelimits.inferenceIpDaily.limit(ip);
+    if (!dailyIpResult.success) {
+      return new Response("Daily IP usage limit exceeded", { status: 429 });
+    }
 
-  const [dailyIpResult, identityResult, globalResult] = await Promise.all([
-    ratelimits.inferenceIpDaily.limit(ip),
-    identity ? ratelimits.inferenceIdentity.limit(identity) : null,
-    ratelimits.inferenceGlobal.limit("all"),
-  ]);
+    const identityResult = identity
+      ? await ratelimits.inferenceIdentity.limit(identity)
+      : null;
+    if (identityResult && !identityResult.success) {
+      return new Response("Daily session usage limit exceeded", {
+        status: 429,
+      });
+    }
 
-  if (!dailyIpResult.success) {
-    return new Response("Daily IP usage limit exceeded", { status: 429 });
-  }
-  if (identityResult && !identityResult.success) {
-    return new Response("Daily session usage limit exceeded", { status: 429 });
-  }
-  if (!globalResult.success) {
-    return new Response("Service usage limit exceeded", { status: 429 });
-  }
+    const globalResult = await ratelimits.inferenceGlobal.limit("all");
+    if (!globalResult.success) {
+      console.error(
+        JSON.stringify({
+          level: "error",
+          message: "global_inference_rate_limit_exceeded",
+          requestId: req.headers.get("x-vercel-id"),
+        }),
+      );
+      return new Response("Service usage limit exceeded", { status: 429 });
+    }
 
-  return null;
+    return null;
+  });
 }
 
 export async function checkAnonymousSessionIssuanceRateLimit(
   req: Request,
 ): Promise<Response | null> {
-  const ratelimits = await ratelimitsPromise;
-  if (!ratelimits) return null;
-
-  const ip = getClientIp(req);
-  const [burstResult, dailyResult] = await Promise.all([
-    ratelimits.sessionIssuanceBurst.limit(ip),
-    ratelimits.sessionIssuanceDaily.limit(ip),
-  ]);
-  if (!burstResult.success || !dailyResult.success) {
-    return new Response("Anonymous session limit exceeded", { status: 429 });
-  }
-  return null;
+  return runRateLimitCheck(req, async (ratelimits) => {
+    const ip = getClientIp(req);
+    const [burstResult, dailyResult] = await Promise.all([
+      ratelimits.sessionIssuanceBurst.limit(ip),
+      ratelimits.sessionIssuanceDaily.limit(ip),
+    ]);
+    if (!burstResult.success || !dailyResult.success) {
+      return new Response("Anonymous session limit exceeded", { status: 429 });
+    }
+    return null;
+  });
 }
 
 export async function checkDownloadRateLimit(
   req: Request,
   identity: string,
 ): Promise<Response | null> {
-  const ratelimits = await ratelimitsPromise;
-  if (!ratelimits) return null;
-
-  const [ipResult, identityResult] = await Promise.all([
-    ratelimits.downloadIp.limit(getClientIp(req)),
-    ratelimits.downloadIdentity.limit(identity),
-  ]);
-  if (!ipResult.success || !identityResult.success) {
-    return new Response("Download rate limit exceeded", { status: 429 });
-  }
-  return null;
+  return runRateLimitCheck(req, async (ratelimits) => {
+    const [ipResult, identityResult] = await Promise.all([
+      ratelimits.downloadIp.limit(getClientIp(req)),
+      ratelimits.downloadIdentity.limit(identity),
+    ]);
+    if (!ipResult.success || !identityResult.success) {
+      return new Response("Download rate limit exceeded", { status: 429 });
+    }
+    return null;
+  });
 }
